Add/Edit Status Report:
- Added pages for and logic for Add and Edit Status Report.
- Added most of the validation for Add and Edit. Will add changes
later if some validations need revising.
- Used textarea for the comments. Are there any other places that we should
use them for? Seems much better. Also we can do validatorMessage if you
guys want to do a custom message. Note that this does not work if youre
validating multiple things.
- Wesley should probably make that look less ugly.
- Wasn't sure what ReportDate was so I currently just set it to current date
so as not to get a null.
- Moved the Add/Edit navigation buttons elsewhere so that it makes more sense.
- Hey, Don, we should probably cut some of the content that you had in the
table and just have it show elsewhere. Like what I did for WP.
- id for Status Report is 'ProjId|WorkPackNo|StatusReportNo' for reference.